+++ CHANGELOG.md
@@ -8,12 +8,14 @@ Add new items at the end of the relevant section under **Unreleased**.
 
 ### Additions
 
-- Add `validate` subcommand to generate validate reports ([#79])
-- Add `--pull-request` option to validate proposals in a pull request ([#86])
+- Add `validate` subcommand to generate validate reports ([#83])
+- Add `--pull-request` option to validate proposals in a pull request ([#87])
 
 ### Changes
 
 - Support extraction of new "Summary of changes" section ([#79])
+- Errors and warnings now include unique error codes ([#90])
+- Add generalized mechanism for validation exemptions ([#92])
 
 ### Fixes
 
@@ -84,4 +86,6 @@ This changelog's format is based on [Keep a Changelog](https://keepachangelog.co
 [#74]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/74
 [#79]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/79
 [#83]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/83
-[#86]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/86
+[#87]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/87
+[#90]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/90
+[#92]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/92