+++ CHANGELOG.md
@@ -16,10 +16,11 @@ Add new items at the end of the relevant section under **Unreleased**.
 - Support extraction of new "Summary of changes" section ([#79])
 - Errors and warnings now include unique error codes ([#90])
 - Add generalized mechanism for validation exemptions ([#92])
-- Promote warnings to errors:
+- Promote warnings to errors ([#93])
     - 'missing review manager'
     - 'missing review manager link'
     - 'missing author link'
+- Remove warning for proposals with past due review period ([#94])
 
 ### Fixes
 
@@ -93,3 +94,5 @@ This changelog's format is based on [Keep a Changelog](https://keepachangelog.co
 [#87]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/87
 [#90]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/90
 [#92]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/92
+[#93]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/93
+[#94]: https://github.com/swiftlang/swift-evolution-metadata-extractor/pull/94